requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
motor>=3.7.0
beanie>=1.27.0
pydantic[email]>=2.12.0
pydantic-settings>=2.7.0
bcrypt>=5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.20
slowapi>=0.1.9
aiosmtplib>=3.0.0
